Calclatorクラス作成
