+++ The-Dentak/Program.cs
@@ -6,26 +6,8 @@ namespace The_Dentak
     {
         static void Main(string[] args)
         {
-            Console.WriteLine("Please input two number.");
-
-            Console.Write("num1:");
-            var firstInputString = Console.ReadLine();
-            if(!int.TryParse(firstInputString, out int num1))
-            {
-                Console.WriteLine($"{firstInputString} is not a number!!");
-                return;
-            }
-            
-            Console.Write("num2:");
-            var secondInputString = Console.ReadLine();
-            if(!int.TryParse(secondInputString, out int num2))
-            {
-                Console.WriteLine($"{secondInputString} is not a number!!");
-                return;
-            }
-
-            var sum = num1 + num2;
-            Console.WriteLine($"{num1} + {num2} = {sum}");
+            var calculator = new Calculator();
+            calculator.Execute();
         }
     }
 }
